Add handling for inaccessible machines
Attempting to get the name or state of an inaccessible machine is not
possible and results in an exception [1]. This commit ensures that
inaccessible machines are omitted from the results. This might be a bit
uninformative, but we cannot even get the name of such machines, so there
is not much we can actually show the user. If users are wondering where
their machines went, they can always open the VirtualBox GUI, which will
show the inaccessible machines along with their name.

[1] VirtualBox Programming Guide and Reference
[http://download.virtualbox.org/virtualbox/SDKRef.pdf], chapter 5,
interface IMachine, attribute accessible